% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sources.R
\name{list_sources}
\alias{list_sources}
\title{List registered political data sources}
\usage{
list_sources()
}
\value{
A character vector of source names.
}
\description{
Returns the names of all data sources registered with the Python \code{catpol}
package (city ordinances, federal laws, executive orders, presidential
speeches, social media archives, etc.).
}
\examples{
\dontrun{
list_sources()
#> [1] "city_san_diego"           "city_san_francisco"
#> [3] "federal_laws"             "federal_executive_orders"
#> [5] "social_trump_truth"       ...
}
}
